+++ elrond-wasm-derive/src/parse/attributes/attr_names.rs
@@ -1,7 +1,9 @@
 pub(super) static ATTR_PAYABLE: &str = "payable";
 pub(super) static ATTR_OUTPUT_NAME: &str = "output_name";
-pub(super) static ATTR_PAYMENT: &str = "payment";
+pub(super) static ATTR_PAYMENT: &str = "payment"; // synonymous with `payment_amount`
+pub(super) static ATTR_PAYMENT_AMOUNT: &str = "payment_amount";
 pub(super) static ATTR_PAYMENT_TOKEN: &str = "payment_token";
+pub(super) static ATTR_PAYMENT_NONCE: &str = "payment_nonce";
 pub(super) static ATTR_VAR_ARGS: &str = "var_args";
 pub(super) static ATTR_INIT: &str = "init";
 pub(super) static ATTR_VIEW: &str = "view";